Completed bottom navigation bar.
