+++ lib/core/navigation/bottom_nav_bar.dart
@@ -1,0 +1,71 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+import 'package:food_delivery_app/src/example_feature/views/cart_page.dart';
+import 'package:food_delivery_app/src/example_feature/views/favorites.dart';
+import 'package:food_delivery_app/src/example_feature/views/home_page.dart';
+import 'package:food_delivery_app/src/example_feature/views/near-me_page.dart';
+import 'package:food_delivery_app/src/example_feature/views/profile_page.dart';
+
+class BottomNavBar extends StatefulWidget {
+  const BottomNavBar({super.key});
+
+  @override
+  State<BottomNavBar> createState() => _BottomNavBarState();
+}
+
+class _BottomNavBarState extends State<BottomNavBar> {
+  int currentIndex = 0;
+
+  final List<Widget> pages = const [
+    HomePage(),
+    NearMePage(),
+    CartPage(),
+    FavoritesPage(),
+    ProfilePage(),
+  ];
+
+  void itemTapped(int index) {
+    setState(() {
+      currentIndex = index;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+        bottomNavigationBar: BottomNavigationBar(
+          showSelectedLabels: false,
+          showUnselectedLabels: false,
+          selectedItemColor: Colors.blueAccent,
+          unselectedItemColor: Colors.grey,
+          currentIndex: currentIndex,
+          type: BottomNavigationBarType.fixed,
+          onTap: itemTapped,
+          items: const [
+            BottomNavigationBarItem(
+              icon: Icon(Icons.home),
+              label: "",
+            ),
+            BottomNavigationBarItem(
+              icon: Icon(Icons.shopping_cart),
+              label: "",
+            ),
+            BottomNavigationBarItem(
+              icon: Icon(Icons.location_searching),
+              label: "",
+              
+            ),
+            BottomNavigationBarItem(
+              icon: Icon(Icons.favorite),
+              label: "",
+            ),
+            BottomNavigationBarItem(
+              icon: Icon(Icons.person),
+              label: "",
+            ),
+          ],
+        ),
+        body: pages[currentIndex],
+      );
+  }
+}
