+++ lib/core/navigation/routes.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:food_delivery_app/src/example_feature/views/login_page.dart';
+
+import '../../src/example_feature/views/auth/login_page.dart';
 
 class Routes {
   static const String loginPage = "loginPage";
