+++ lib/core/services/auth/auth_gate.dart
@@ -1,8 +1,9 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import 'package:food_delivery_app/core/navigation/bottom_nav_bar.dart';
 
 import '../../../src/example_feature/views/home_page.dart';
-import '../../../src/example_feature/views/login_page.dart';
+import '../../../src/example_feature/views/auth/login_page.dart';
 
 class AuthGate extends StatelessWidget {
   const AuthGate({super.key});
@@ -16,7 +17,7 @@ class AuthGate extends StatelessWidget {
         stream: FirebaseAuth.instance.authStateChanges(),
         builder: (context, snapshot) {
           if (snapshot.hasData) {
-            return const HomePage();
+            return const BottomNavBar();
           } else if (snapshot.connectionState == ConnectionState.waiting) {
             return const CircularProgressIndicator();
           }
