+++ lib/src/example_feature/views/cart_page.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class CartPage extends StatefulWidget {
+  const CartPage({super.key});
+
+  @override
+  State<CartPage> createState() => _CartPageState();
+}
+
+class _CartPageState extends State<CartPage> {
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text("Cart Page"),
+      ),
+    );
+  }
+}
