+++ lib/src/example_feature/views/favorites.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class FavoritesPage extends StatefulWidget {
+  const FavoritesPage({super.key});
+
+  @override
+  State<FavoritesPage> createState() => _FavoritesPageState();
+}
+
+class _FavoritesPageState extends State<FavoritesPage> {
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text("Favorites Page"),
+      ),
+    );
+  }
+}
