+++ lib/src/example_feature/views/near-me_page.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class NearMePage extends StatefulWidget {
+  const NearMePage({super.key});
+
+  @override
+  State<NearMePage> createState() => _NearMePageState();
+}
+
+class _NearMePageState extends State<NearMePage> {
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text("Near-Me Page"),
+      ),
+    );
+  }
+}
